mcp/servers/chanel_knowledge_server/pyproject.toml
[project]
name = "chanel-knowledge-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.11.0",
    "azure-search-documents>=11.5.3",
    "azure-identity>=1.23.1",
    "openai>=1.96.1"
]
